fix some logic and fix unit test
